refactor: 对话框外观逻辑重构与交互行为开关修复
- 新增 src/utils/color.ts: 提取颜色工具函数
- 新增 src/composables/useDialogAppearance.ts: 封装外观样式和交互行为
- 新增 src/components/settings/dialog/DialogAppearancePanel.vue: 独立设置面板
- 重构 GameDialog.vue: 使用 useDialogAppearance composable
- 重构 SettingsBackground.vue: 使用 DialogAppearancePanel 组件
- 修复: 使用 computed() 包装 store getter 确保响应式更新
- 修复: 移除不存在的 'idle' 状态，仅保留 'thinking' 和 'responding'
